+++ ui/src/components/Icons.tsx
@@ -45,6 +45,7 @@ export const Icon = {
   Bell:      (p: Props) => <Base {...p}><path d="M6 8a6 6 0 1 1 12 0c0 7 3 8 3 8H3s3-1 3-8z"/><path d="M10 21a2 2 0 0 0 4 0"/></Base>,
   Users:     (p: Props) => <Base {...p}><circle cx="9" cy="8" r="3.5"/><circle cx="17" cy="9" r="2.5"/><path d="M2.5 21c0-3.6 2.9-6.5 6.5-6.5s6.5 2.9 6.5 6.5"/><path d="M15 21c0-2.6 1.6-4.9 4-5.8"/></Base>,
   Trash:     (p: Props) => <Base {...p}><path d="M4 7h16M10 11v6M14 11v6M5 7l1 13a2 2 0 0 0 2 2h8a2 2 0 0 0 2-2l1-13M9 7V4a1 1 0 0 1 1-1h4a1 1 0 0 1 1 1v3"/></Base>,
+  Download:  (p: Props) => <Base {...p}><path d="M12 4v12M7 11l5 5 5-5M4 20h16"/></Base>,
   Refresh:   (p: Props) => <Base {...p}><path d="M3 12a9 9 0 1 0 3-6.7L3 8M3 3v5h5"/></Base>,
   Warning:   (p: Props) => <Base {...p}><path d="M12 3l10 18H2L12 3z"/><path d="M12 10v5M12 18v.5"/></Base>,
   Link:      (p: Props) => <Base {...p}><path d="M10 14a4 4 0 0 0 5.7 0l3-3a4 4 0 0 0-5.7-5.7l-1 1"/><path d="M14 10a4 4 0 0 0-5.7 0l-3 3a4 4 0 0 0 5.7 5.7l1-1"/></Base>,